+++ .changeset/bright-phones-serve.md
@@ -1,0 +1,5 @@
+---
+"@yamada-ui/utils": patch
+---
+
+Updated import paths for better clarity.
